requirements.txt
requests
aiohttp
pycoingecko
python-dotenv
flask